requirements.txt
nicegui>=1.4.0
requests>=2.31.0

# Authentication
passlib[bcrypt]
pyrebase4>=4.7.1
firebase-admin>=6.2.0

# Database
psycopg2-binary>=2.9.9

# Environment variables
python-dotenv>=1.0.0

# Async support
aiohttp>=3.8.5

# Date/time handling
python-dateutil>=2.8.2
pytz>=2023.3

# JSON handling
ujson>=5.8.0
pandas>=2.1.0

# Data visualization
wordcloud>=1.8.1
nltk>=3.8.1
plotly>=6.0.1

# AI and Language Models
openai>=1.6.0
langchain>=0.3.15
langchain-openai>=0.2.14
langchain-anthropic>=0.3.5
langchain-core>=0.3.33
pydantic>=2.0.0
markdown>=3.4.1
